requirements.txt
openai~=1.35.14

requests~=2.31.0
python-dotenv~=1.0.1
groq~=0.10.0